Add js-resume_schematic-wires.xd to wireframes directory
The artboards with schematic wireframes (of diagrams and site pages per artboard) as presented on 03 Sep 2020 for the project's initial mentor meeting are arranged in an Adobe XD file previously named JS-REDUX-02.xd. Excess and draft artboards in JS-REDUX-02.xd have been removed in order to freeze and package the initial set of wireframes as a Schematic-Set. This is now named js-resume_schematic-wires.xd which contains 1 expansive artboard of diagrams and 22+ artboards of wireframes for web and mobile. The document js-resume_schematic-wires will be saved in wireframes/schematic along with exported pdfs/pngs of its independent content. Forthcoming development of sketches and wireframes produced after 03 Sep 2020 shall be saved in wireframes/development.