Respecify global section configs for efficient maintenance
Article > section selectors were replaced with direct section selectors. Sections collapseZero to collapseSix have to be direct children of data-parent='resumeStack' for the .show event to scroll to the top of the toggled collapse-section. Sections will have to be unnested from article-parents in index.html for the inital release, until custom JavaScript is written to modify the .collapse.collapsing transitions and events. Global configs to sections and per-section variables are now contained and maintained to look the same regardless.