Refactor class nomenclature and css ruleset composition
In addition to improving code quality, comments thoroughly explain components and ruleset blocks. Style assignments are now to class names instead of html elements for interchangeability in the future. Development of side menu and scroll spy along with revised animation for the nav element. 80% of the css has been prefixed. Overall collapse transitions now animate flexbox to achieve a smooth scroll-to-top effect.